+++ src/main/java/com/drathonix/loadmychunks/common/config/regis/LoadStateRetriever.java
@@ -1,0 +1,27 @@
+package com.drathonix.loadmychunks.common.config.regis;
+
+import com.drathonix.loadmychunks.common.registry.custom.LoadStateRegistry;
+import com.drathonix.loadmychunks.common.system.control.ILoadState;
+import com.vicious.persist.mappify.registry.Stringify;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Item;
+
+public class LoadStateRetriever extends RegistryRetriever<ILoadState> {
+    static {
+        Stringify.register(LoadStateRetriever.class, LoadStateRetriever::new, LoadStateRetriever::serializable);
+    }
+
+    public LoadStateRetriever(String key) {
+        super(key);
+    }
+
+    public LoadStateRetriever(ResourceLocation key) {
+        super(key);
+    }
+
+    @Override
+    ILoadState retrieve() {
+        return LoadStateRegistry.INSTANCE.get(location);
+    }
+}